oi_reports/pyproject.toml
[tool.poetry]
name = "oi_reports"
version = "0.1.0"
description = ""
authors = ["Pranav Dua <pdua@twilio.com>"]

[tool.poetry.dependencies]
python = "^3.8"
google-auth = "^2.6.0"
APScheduler = "^3.9.1"
google-oauth = "^1.0.0"
google-api-python-client-stubs = "^1.11.0"
google-auth-oauthlib = "^0.5.1"
google-auth-httplib2 = "^0.1.0"
coverage = "^6.4.2"
pandas = "^1.5.0"
pyarrow = "^9.0.0"
datadog = "^0.44.0"
datadog-api-client = "^2.8.0"
Flask = "^2.2.3"
google-cloud-bigquery = "^3.7.0"
Flask-RESTful = "^0.3.9"

[tool.poetry.dev-dependencies]
isort = "5.6.4"
black = "20.8b1"
pyflakes = "2.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
